capstone/requirements.txt
absl-py==0.7.1
appnope==0.1.0
astor==0.7.1
attrs==19.1.0
backcall==0.1.0
bert-serving-client==1.8.9
bert-serving-server==1.8.9
bleach==3.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
dash==0.42.0
dash-core-components==0.47.0
dash-daq==0.1.0
dash-html-components==0.16.0
dash-renderer==0.23.0
dash-table==3.6.0
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
Flask==1.0.2
Flask-Compress==1.4.0
gast==0.2.2
GPUtil==1.4.0
grpcio==1.16.1
h5py==2.9.0
idna==2.8
ipykernel==5.1.0
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.1
MarkupSafe==1.1.1
mistune==0.8.4
mkl-fft
mkl-random
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.3
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
patsy==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
plotly==3.9.0
plotly-express==0.2.2
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.1
ptyprocess==0.6.0
Pygments==2.4.0
pyrsistent==0.15.1
python-dateutil==2.8.0
pytz==2019.1
pyzmq==18.0.1
qtconsole==4.4.4
requests==2.21.0
retrying==1.3.3
scipy==1.2.1
Send2Trash==1.5.0
six==1.12.0
statsmodels==0.9.0
tensorboard==1.11.0
tensorflow==1.11.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
urllib3==1.24.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.2
widgetsnbextension==3.4.2
xlrd==1.2.0
